# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "PopulationSnapshot": {
        "properties": {
          "data_kind": {
            "const": "synthetic",
            "title": "Data Kind",
            "type": "string"
          },
          "household_reference": {
            "format": "date",
            "title": "Household Reference",
            "type": "string"
          },
          "households": {
            "title": "Households",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_fixture": {
            "title": "Is Fixture",
            "type": "boolean"
          },
          "located_persons": {
            "title": "Located Persons",
            "type": "integer"
          },
          "manifest_sha256": {
            "title": "Manifest Sha256",
            "type": "string"
          },
          "municipalities": {
            "title": "Municipalities",
            "type": "integer"
          },
          "persons": {
            "title": "Persons",
            "type": "integer"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "reference_date": {
            "format": "date",
            "title": "Reference Date",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "manifest_sha256",
          "reference_date",
          "household_reference",
          "persons",
          "households",
          "municipalities",
          "located_persons",
          "is_fixture",
          "data_kind",
          "published_at"
        ],
        "title": "PopulationSnapshot",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Verified synthetic population snapshots and their statistical evidence. No real-person records. Only published releases are visible. Demo data are explicitly labelled. Values are decimal strings; missing and suppressed values are null.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Itadb \u2014 popolazione sintetica",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v3/populations": {
      "get": {
        "operationId": "snapshots_v3_populations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PopulationSnapshot"
                  },
                  "title": "Response Snapshots V3 Populations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Snapshots",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}": {
      "get": {
        "operationId": "snapshot_v3_populations__snapshot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PopulationEvidence"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Snapshot",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}/comparison": {
      "get": {
        "description": "Observed input constraints and corresponding counts from imported synthetic records.",
        "operationId": "comparison_v3_populations__snapshot_id__comparison_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "municipality_code",
            "required": true,
            "schema": {
              "pattern": "^[0-9]{6}$",
              "title": "Municipality Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "sex_age",
              "enum": [
                "sex_age",
                "foreign_age",
                "citizenship",
                "household_size"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PopulationComparison"
                  },
                  "title": "Response Comparison V3 Populations  Snapshot Id  Comparison Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Comparison",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}/distributions": {
      "get": {
        "operationId": "distributions_v3_populations__snapshot_id__distributions_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "municipality_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9]{6}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Municipality Code"
            }
          },
          {
            "in": "query",
            "name": "region_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9]{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Code"
            }
          },
          {
            "in": "query",
            "name": "sex",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "M",
                    "F"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sex"
            }
          },
          {
            "in": "query",
            "name": "citizenship_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Citizenship Code"
            }
          },
          {
            "in": "query",
            "name": "age_min",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100,
              "minimum": 0,
              "title": "Age Min",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "age_max",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 0,
              "title": "Age Max",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PopulationDistribution"
                  },
                  "title": "Response Distributions V3 Populations  Snapshot Id  Distributions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Distributions",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}/households": {
      "get": {
        "operationId": "households_v3_populations__snapshot_id__households_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "municipality_code",
            "required": true,
            "schema": {
              "pattern": "^[0-9]{6}$",
              "title": "Municipality Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 6,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Size"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "household_id",
              "enum": [
                "household_id",
                "size"
              ],
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 70000000,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyntheticHouseholdPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Households",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}/households/{household_id}": {
      "get": {
        "operationId": "household_v3_populations__snapshot_id__households__household_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "maximum": 70000000,
              "minimum": 1,
              "title": "Household Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyntheticHouseholdDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Household",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}/map": {
      "get": {
        "description": "Representative municipality locations, never synthetic individual residences.",
        "operationId": "population_map_v3_populations__snapshot_id__map_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9]{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Code"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PopulationMap"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Population Map",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}/municipalities": {
      "get": {
        "operationId": "municipalities_v3_populations__snapshot_id__municipalities_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9]{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Code"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 999999,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PopulationMunicipalityPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Municipalities",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}/persons": {
      "get": {
        "operationId": "persons_v3_populations__snapshot_id__persons_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "municipality_code",
            "required": true,
            "schema": {
              "pattern": "^[0-9]{6}$",
              "title": "Municipality Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sex",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "M",
                    "F"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sex"
            }
          },
          {
            "in": "query",
            "name": "citizenship_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Citizenship Code"
            }
          },
          {
            "in": "query",
            "name": "age_min",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100,
              "minimum": 0,
              "title": "Age Min",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "age_max",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 0,
              "title": "Age Max",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "person_id",
              "enum": [
                "person_id",
                "age",
                "sex",
                "citizenship_code",
                "household_id"
              ],
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 70000000,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyntheticPersonPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Persons",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
    "/v3/populations/{snapshot_id}/persons/{person_id}": {
      "get": {
        "operationId": "person_v3_populations__snapshot_id__persons__person_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "maximum": 70000000,
              "minimum": 1,
              "title": "Person Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyntheticPerson"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Person",
        "tags": [
          "Popolazione sintetica"
        ]
      }
    },
...
  }
}